pyproject.toml
[project]
name = "yun-recs"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.10.2",
    "python-frontmatter>=1.1.0",
]
